refactor: move mux to root folder
